requirements.txt
Flask
requests
python-dotenv
langchain
langchain-openai
langchain-community
langchain-huggingface
psycopg2-binary
sqlalchemy
faiss-cpu
psutil